requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
deepface
tf-keras
numpy
qrcode
pillow
opencv-python-headless
